+++ src/components/ContactItem.jsx
@@ -1,9 +1,23 @@
-import { Paper, Avatar, IconButton, Typography, Box } from '@mui/material';
+import * as React from 'react';
+import {
+  Paper,
+  Avatar,
+  IconButton,
+  Typography,
+  Box,
+  Dialog,
+  Slide,
+} from '@mui/material';
+import CreateIcon from '@mui/icons-material/Create';
+import CloseIcon from '@mui/icons-material/Close';
 import DeleteIcon from '@mui/icons-material/Delete';
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
 import FavoriteIcon from '@mui/icons-material/Favorite';
-import EditContactWindow from './EditContactWindow';
+import EditContactForm from './EditContactForm';
 
+const Transition = React.forwardRef(function Transition(props, ref) {
+  return <Slide direction="up" ref={ref} {...props} />;
+});
 export default function ContactItem({
   id,
   name,
@@ -13,6 +27,15 @@ export default function ContactItem({
   favorite,
   onToggleFavoriteContact,
 }) {
+  const [open, setOpen] = React.useState(false);
+
+  const handleClickOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
   const firstLetter = name[0];
   return (
     <Paper sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
@@ -50,7 +73,33 @@ export default function ContactItem({
         >
           <DeleteIcon fontSize="inherit" />
         </IconButton>
-        <EditContactWindow id={id} />
+        <div>
+          <IconButton
+            size="small"
+            color="primary"
+            aria-label="edit"
+            onClick={handleClickOpen}
+          >
+            <CreateIcon fontSize="inherit" />
+          </IconButton>
+
+          <Dialog
+            open={open}
+            onClose={handleClose}
+            TransitionComponent={Transition}
+          >
+            <IconButton
+              sx={{ width: '25px', height: '25px' }}
+              color="inherit"
+              onClick={handleClose}
+              aria-label="close"
+            >
+              <CloseIcon />
+            </IconButton>
+
+            <EditContactForm id={id} />
+          </Dialog>
+        </div>
         <IconButton
           size="small"
           aria-label="favorite"
